Google Gemini/Multilingual Invoice Insight Harnessing Google's Generative AI for Effortless Data Extraction/Code/requirements.txt
streamlit
google-generativeai
langchain
python-dotenv
PyPDF2 
spacy           # python -m spacy download en_core_web_sm  ."Run in Terminal"
googletrans
pandas